+++ .gitignore
@@ -7,3 +7,5 @@
 .env
 __pycache__
 .ipynb_checkpoints
+/court_scrape/*
+!/court_scrape/lambda_function.py
