+++ Makefile
@@ -1,3 +1,10 @@
-ec2-connect
-	ssh -i ~/.ssh/cuyacourts.pem -L 8888:localhost:8888 ec2-user@ec2-44-210-91-132.compute-1.amazonaws.com
+init-env:
+	brew install pyenv pyenv-virtualenv
+	eval "$(pyenv init-)"
+	eval "$(pyenv virtualenv-init -)";  
+	export PATH="$HOME/.pyenv/bin:$PATH"
+	pyenv install 3.9
+	pyenv virtualenv 3.9 cuya-courts
 
+env:
+	pyenv local cuya-courts
