+++ requirements.txt
@@ -1,7 +1,9 @@
-jupyter==1.0.0
 pandas==1.3.5
 SQLAlchemy==1.4.40
 chromedriver-py==105.0.5195.19
 selenium==4.4.2
 webdriver-manager==3.8.3
 psycopg2==2.9.3
+
+# brew install aws-sam-cli
+#
